example-requirements.txt
metaflow
scikit-learn
plotly-express
kaleido
pandas
numpy
scipy
jupyterlab
kaggle
